Anitube: Fix player and genres list + refactor
